+++ libs/ngx-slice-kit/src/public-api.ts
@@ -19,7 +19,16 @@ export { ThemeModule } from './lib/core/theme/theme.module';
 export { ThemeService } from './lib/core/theme/theme.service';
 export { ThemeDirective } from './lib/core/theme/theme.directive';
 export {
-    Theme, ColorProperty, ThemeBaseColors, ThemeColors, MixinRgba, NumArrayToRgbString, RgbaToHex,
+    // models
+    Theme,
+    ColorProperty,
+    ThemeBaseColors,
+    ThemeColors,
+    // methods
+    MixinRgba,
+    NumArrayToRgbString,
+    RgbaToHex,
+    HexToRgb
 } from './lib/core/theme/theme.model';
 export { themeDark } from './lib/core/theme/lib/theme-dark';
 export { themeLight } from './lib/core/theme/lib/theme-light';