+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 web:
 - pagination component demo
 - sidenav component demo
+- demo component stackblitz url
 
 ### Changed
 web: