+++ CHANGELOG.md
@@ -23,6 +23,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 web:
 - enhanced theme palette guide
 
+ngx-slice-kit:
+- HexToRgba theme method
+
 ### Changed
 ngx-slice-kit:
 - theme model color handling updated